Server/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
python-decouple